pre-pull

+++ plugins/epan/etsi_ieee1609dot2/packet-etsi_ieee1609dot2.c
@@ -690,7 +690,7 @@ dissect_ieee1609dot2_appPermissions_packet(tvbuff_t *tvb, packet_info *pinfo _U_
   printf(">>> dissect_ieee1609dot2_appPermissions_packet: offset=0x%02x\n", offset);
   if (tree) { /* we are being asked for details */
     gint sh_len;
-    int items = 0;
+    guint8 items = 0;
     guint8 len = 0;
     
     /* Sec Header tree - See IEEE Std 1609.2a-2017 */
